pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "naturalization-quiz"
version = "0.1.0"
description = "A CLI quiz for US naturalization test questions"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pyyaml"
]
authors = [
    {name = "Edoardo Balzani", email = "ebalzani87@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
]

[project.scripts]
naturalization-quiz = "naturalization_quiz.quiz:main"

[project.urls]
Homepage = "https://github.com/BalzaniEdoardo/naturalization_quiz"

[tool.setuptools.package-data]
naturalization_quiz = ["*.yaml"]